Replaces integer(1) by integer(int8)

+++ examples/cairo-tests.f90
@@ -23,7 +23,7 @@
 !------------------------------------------------------------------------------
 ! Contributed by Jerry DeLisle and Vincent Magnin
 ! GTK 4 version: vmagnin 2020-05-19
-! Last modification: vmagnin 2020-05-28, 2021-06-07
+! Last modification: vmagnin 2020-05-28, 2022-04-05
 !------------------------------------------------------------------------------
 
 module handlers
@@ -176,13 +176,14 @@ contains
   subroutine Mandelbrot_set(my_drawing_area, xmin, xmax, ymin, ymax, itermax)
     ! Whole set: xmin=-2d0, xmax=+1d0, ymin=-1.5d0, ymax=+1.5d0, itermax=1000
     ! Seahorse valley:  around x=-0.743643887037151, y=+0.13182590420533, itermax=5000
+    use, intrinsic :: iso_fortran_env, only: wp=>real64, int8
 
     type(c_ptr) :: my_drawing_area
     integer(4) :: i, j, k, p, itermax
     real(8)    :: x, y, xmin, xmax, ymin, ymax ! coordinates in the complex plane
     complex(8) :: c, z
     real(8)    :: scx, scy             ! scales
-    integer(1) :: red, green, blue     ! rgb color
+    integer(int8) :: red, green, blue     ! rgb color
     real(8) :: t0, t1
 
     print *, "Entering Mandelbrot_set() subroutine"
