+++ examples/julia_pixbuf.f90
@@ -22,7 +22,7 @@
 ! If not, see <http://www.gnu.org/licenses/>.
 !------------------------------------------------------------------------------
 ! Contributed by Vincent Magnin and Jerry DeLisle
-! Last modifications: vmagnin+Ian Harvey 2019-02-21, vmagnin 2021-06-07
+! Last modifications: vmagnin+Ian Harvey 2019-02-21, vmagnin 2022-04-05
 !------------------------------------------------------------------------------
 
 module global_widgets
@@ -407,7 +407,7 @@ contains
   ! The scientific computing is done here
   !*********************************************
   subroutine Julia_set(xmin, xmax, ymin, ymax, c, itermax)
-    use, intrinsic :: iso_fortran_env, only: wp=>real64
+    use, intrinsic :: iso_fortran_env, only: wp=>real64, int8
     use, intrinsic :: iso_c_binding
     use global_widgets
 
@@ -415,7 +415,7 @@ contains
     real(wp)    :: x, y, xmin, xmax, ymin, ymax ! coordinates in the complex plane
     complex(wp) :: c, z
     real(wp)    :: scx, scy       ! scales
-    integer(1)  :: red, green, blue     ! rgb color
+    integer(int8)  :: red, green, blue     ! rgb color
     real(wp)    :: t0, t1
 
     computing = .true.
